Final version ✨

+++ src/main/java/org/esprim/tp1/repository/FoyerRepository.java
@@ -3,5 +3,5 @@ package org.esprim.tp1.repository;
 import org.esprim.tp1.entity.Foyer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FoyerRepository extends JpaRepository<Foyer, Integer> {
+public interface FoyerRepository extends JpaRepository<Foyer, Long> {
 }
