+++ src/main/java/org/esprim/tp1/repository/UniversiteRepository.java
@@ -3,5 +3,5 @@ package org.esprim.tp1.repository;
 import org.esprim.tp1.entity.Universite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UniversiteRepository extends JpaRepository<Universite, Integer> {
+public interface UniversiteRepository extends JpaRepository<Universite, Long> {
 }
